requirements.txt
alembic==1.11.1
bcrypt==4.0.1
cryptography==40.0.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-SSLify==0.1.5
flask-swagger-ui==4.11.1
pyjwt==2.7.0
PyMySQL==1.1.0
SQLAlchemy==2.0.17
Werkzeug==2.3.6
nltk==3.8.1
python-dotenv==1.0.0
